requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.3
django-ninja==0.22.2
django-ninja-auth==0.1.7
djangorestframework==3.15.0
idna==3.6
Naked==0.1.32
numpy==1.26.4
pandas==2.2.1
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
